Support fields
This supports fields that can be seen in the `list` commands in wide
view, `dump` and `show` commands, as well as the `edit` commands.

There is nothing that can be done with them yet, they are only for
decoration. Fields for remote databases are saved and maintained
locally, they are not persisted to the remote resource.

Signed-off-by: Erik Hollensbe <git@hollensbe.org>