Remove sync window (unused), add query window setting to saturn config
This adds the query window setting and removes sync window, which was
supposed to supply this (and I promptly forgot about it, so it was dead
code). There's still a stub left to avoid breaking existing
configuration files, but it'll immediately rewrite them to not have the
item in there.

Signed-off-by: Erik Hollensbe <git@hollensbe.org>